[FIX] logfile for Action daemon
The daemonize module for Tapper::Action creates its own log file which
mainly (only?) contains daemon started/stopped. This file needs to be
writable both during tests and in production. Therefore, it can not live
in /var/log which is writable only by root per default. Having it in
/tmp/ is still based on an assumption (that /tmp/ is writable by
everyone) but better than the version before.