[FIX] load classes without eval.
When the devil invented the string eval he only slightly hid the evil
within it by changing the "i" to "a". We certainly don't want user to be
able to execute arbitrary code by giving plugin names like "Foo;
system('rm -rf /')". So instead of selling our souls we now use plugin
classes with Class::Load. This module also protects agains the double
require bug in perl. If you require a module and it fails for any other
reason than "file does not exist" any subsequent requires of the same
module will be reported as success because perl only checks is internal
data structure on whether it ever tried to load the module. Class::Load
can protect against this issue.