fix test config
For testing the reset plugin we let Tapper::Action copy a file to a temp
directory. To make sure the correct directory is chosen, we use a
special config file. This config file is merged with the default config
file and only then is the environment (live, devel, test)
applied. Therefore, the test config file needs all entries below test:
The default config and the test config had the same values for
grubpath. Therefore, the issue was not detected until recently, when I
used a different default config file.